+++ test/stream.cc
@@ -8,17 +8,17 @@ TEST(md5sum, stream) {
     const auto test_data = build_test_data(256 * 256);
 
     MD5 md5;
-    for (uint32_t size = 1; size <= 256; ++size) {
+    for (size_t size = 1; size <= 256; ++size) {
         auto expect = MD5::Hash(test_data.data(), size * 256);
 
-        for (int times = 0; times < 256; ++times) {
+        for (size_t times = 0; times < 256; ++times) {
             const auto offset = test_data.data() + times * size;
             md5.Update(offset, size); // update multiple times
         }
         EXPECT_EQ(md5.Final().Digest(), expect);
         md5.Reset(); // reset for next round
 
-        for (int times = 0; times < 256; ++times) {
+        for (size_t times = 0; times < 256; ++times) {
             const auto offset = test_data.data() + times * size;
             md5.Update(std::string_view {offset, size}); // update multiple times
         }
